cmd/cue: use module information if no -X version information specified
Currently goreleaser is used to build and publish binary artefacts to
GitHub for each release. As part of this process, cmd/cue is built with
cuelang.org/cmd/cue as the main module, and a build flag of:

    -ldflags='-X main.version=<version>'

to inject the release version.

But when cmd/cue is resolved as a module requirement (via
proxy.golang.org) and installed, no version information is available; as
such 'cue version' simply shows 'custom'. This is the case when
installing cue as a tool dependency of a project or via gobin
(github.com/myitcv/gobin).

I have raised goreleaser/goreleaser#1354 to request that goreleaser add
support for building artefacts based on proxy.golang.org-resolved
modules, with module information for the main package's module baked in.
This will ensure consistency between published release artefacts and
anyone install cmd/cue as a tool dependency, for example.  But until
that support lands, we need to continue to support the existing ldflags
version injection approach.

Hence if there is no version information injected we fallback to using
module information if it is available.